# go/go/src/ipsap/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "IPSAP Service APIs",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "basePath": "/api/v1.0",
    "paths": {
        "/admin/user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "사용자 관리 목록",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin(행정간사)"
                ],
                "summary": "사용자 관리 목록",
                "responses": {
                    "200": {}
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "사용자 관리 User 등록",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin(행정간사)"
                ],
                "summary": "사용자 관리 User 등록",
                "parameters": [
                    {
                        "description": "param",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.userBasicModel"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "행정간사가 사용자 정보 수정",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin(행정간사)"
                ],
                "summary": "행정간사가 사용자 정보 수정",
                "parameters": [
                    {
                        "description": "param",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.adminPatchUserInfoModel"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/admin/user-batch": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "사용자 관리 User 일괄 등록",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin(행정간사)"
                ],
                "summary": "사용자 관리 User 일괄 등록",
                "parameters": [
                    {
                        "description": "param",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.userBatchModel"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/admin/user/{user_seq}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "등록 대기중인 사용자 삭제",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin(행정간사)"
                ],
                "summary": "등록 대기중인 사용자 삭제",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_seq",
                        "name": "user_seq",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/admin/user/{user_seq}/resend-msg": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "사용자 관리 등록 대기 상태인 유저 가입 메세지 재발송",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin(행정간사)"
                ],
                "summary": "사용자 관리 등록 대기 상태인 유저 가입 메세지 재발송",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_seq",
                        "name": "user_seq",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/admin/user/{user_seq}/reset-password": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "사용자 관리 비밀번호 초기화",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin(행정간사)"
                ],
                "summary": "사용자 관리 비밀번호 초기화",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_seq",
                        "name": "user_seq",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/admin/user/{user_seq}/withdraw": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "회원 강제 탈퇴(실제 삭제 X 상태만 변경)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin(행정간사)"
                ],
                "summary": "회원 강제 탈퇴",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_seq",
                        "name": "user_seq",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
...
    },
    "definitions": {
...
        "model.adminPatchUserInfoModel": {
            "type": "object",
            "properties": {
                "agree_email": {
                    "type": "integer",
                    "example": 1
                },
                "agree_pri_open": {
                    "type": "integer",
                    "example": 1
                },
                "agree_sms": {
                    "type": "integer",
                    "example": 1
                },
                "dept": {
                    "type": "string",
                    "example": "소속부서"
                },
                "edu_course_num": {
                    "type": "string",
                    "example": "0"
                },
                "edu_date": {
                    "type": "string",
                    "example": "20210101"
                },
                "edu_institution": {
                    "type": "string",
                    "example": "건국대학교"
                },
                "major_field": {
                    "type": "string",
                    "example": "수의학"
                },
                "name": {
                    "type": "string",
                    "example": "홍길동"
                },
                "phoneno": {
                    "type": "string",
                    "example": "021231234"
                },
                "position": {
                    "type": "string",
                    "example": "직급"
                },
                "user_seq": {
                    "type": "integer",
                    "example": 1
                },
                "user_type": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
...
        "model.userBasicModel": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "test@test.com"
                },
                "phoneno": {
                    "type": "string",
                    "example": "01012341234"
                },
                "user_type": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
        "model.userBatchModel": {
            "type": "object",
            "properties": {
                "userArr": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.userBasicModel"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}